requirements.txt
torch>=2.0.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.4
PyYAML==6.0.1
tensorboard==2.15.2
pytest==7.4.4
